+++ R/DBA.R
@@ -26,6 +26,8 @@
 #'
 #' # Change the provided order
 #' dtw.avg2 <- DBA(CharTraj[5:1], CharTraj[[1]], trace = TRUE)
+#'
+#' # Same result?
 #' all(dtw.avg == dtw.avg2)
 #'
 #' @param X A data matrix where each row is a time series. Optionally, a list where each element is a time series.