+++ R/dtw-lb.R
@@ -15,6 +15,14 @@
 #' between \eqn{x_{i-10}} and \eqn{x_{i+10}}, resulting in \code{10*2 + 1 = 21} observations falling within
 #' the window.
 #'
+#' @section Parallel Computing:
+#'
+#' Please note that running tasks in parallel does \strong{not} guarantee faster computations.
+#' The overhead introduced is sometimes too large, and it's better to run tasks sequentially.
+#'
+#' The user can register a parallel backend with the \code{doParallel} package in order to attempt to
+#' speed up the calculations (see the examples).
+#'
 #' @note
 #'
 #' This function uses a lower bound that is only defined for time series of equal lengths.
@@ -38,7 +46,7 @@
 #' data(uciCT)
 #'
 #' # Reinterpolate to same length
-#' data <- lapply(CharTraj, reinterpolate, newLength = 205)
+#' data <- lapply(CharTraj, reinterpolate, newLength = 180)
 #'
 #' # Calculate the DTW distance between a certain subset aided with the lower bound
 #' system.time(d <- dtw_lb(data[1:5], data[6:50], window.size = 20))
@@ -55,6 +63,29 @@
 #'
 #' # Same results?
 #' all(NN1 == NN2)
+#'
+#' \dontrun{
+#' #### Running DTW_LB with parallel support
+#' # For such a small dataset, this is probably slower in parallel
+#' require(doParallel)
+#'
+#' # Create parallel workers
+#' cl <- makeCluster(detectCores())
+#' registerDoParallel(cl)
+#'
+#' # Distance matrix
+#' D <- dtw_lb(data[1:50], data[51:100], window.size = 20)
+#'
+#' # Stop parallel workers
+#' stopCluster(cl)
+#'
+#' # Return to sequential computations
+#' registerDoSEQ()
+#'
+#' # Nearest neighbors
+#' NN <- apply(D, 1, which.min)
+#' cbind(names(data[1:50]), names(data[51:100][NN]))
+#' }
 #'
 #' @author Alexis Sarda-Espinosa
 #'
@@ -90,6 +121,14 @@ dtw_lb <- function(x, y = NULL, window.size = NULL, norm = "L1", error.check = T
                       error.check = error.check,
                       ...)
 
+     ## Attempt parallel computations?
+     do_par <- foreach::getDoParRegistered()
+
+     if (do_par)
+          workers <- foreach::getDoParWorkers()
+     else
+          workers <- 1L
+
      ## For indexing convenience
      d <- t(d)
      singleIndexing <- seq(from=0, by=nrow(d), length.out=ncol(d))
@@ -100,21 +139,61 @@ dtw_lb <- function(x, y = NULL, window.size = NULL, norm = "L1", error.check = T
      indNN <- new.indNN + 1
 
      while (any(new.indNN != indNN)) {
-          indNew <- new.indNN != indNN
+          indNew <- which(new.indNN != indNN)
           indNN <- new.indNN
 
-          dSub <- switch(EXPR = norm,
+          if (do_par) {
+               tasks <- parallel::splitIndices(length(indNew), workers)
 
-                         L1 = proxy::dist(X[indNew], Y[indNN[indNew]], pairwise = TRUE,
-                                          method = "DTW", dist.method = "L1",
-                                          window.type = "slantedband", window.size = window.size),
+               indNew <- lapply(tasks, function(id) {
+                    indNew[id]
+               })
 
-                         L2 = proxy::dist(X[indNew], Y[indNN[indNew]], pairwise = TRUE,
-                                          method = "DTW2",
-                                          window.type = "slantedband", window.size = window.size))
+               exclude <- setdiff(ls(), c("X", "Y", "norm", "indNN", "window.size"))
 
-          indD <- indNN + singleIndexing
-          d[indD[indNew]] <- dSub
+               dSub <- foreach(indNew = indNew,
+                               .combine = c,
+                               .multicombine = TRUE,
+                               .packages = "dtwclust",
+                               .noexport = exclude) %dopar% {
+                                    switch(EXPR = norm,
+
+                                           L1 = proxy::dist(X[indNew], Y[indNN[indNew]],
+                                                            pairwise = TRUE,
+                                                            method = "DTW",
+                                                            dist.method = "L1",
+                                                            window.type = "slantedband",
+                                                            window.size = window.size),
+
+                                           L2 = proxy::dist(X[indNew], Y[indNN[indNew]],
+                                                            pairwise = TRUE,
+                                                            method = "DTW2",
+                                                            window.type = "slantedband",
+                                                            window.size = window.size))
+                               }
+
+               indD <- indNN + singleIndexing
+               d[indD[unlist(indNew)]] <- dSub
+
+          } else {
+               dSub <- switch(EXPR = norm,
+
+                              L1 = proxy::dist(X[indNew], Y[indNN[indNew]],
+                                               pairwise = TRUE,
+                                               method = "DTW",
+                                               dist.method = "L1",
+                                               window.type = "slantedband",
+                                               window.size = window.size),
+
+                              L2 = proxy::dist(X[indNew], Y[indNN[indNew]],
+                                               pairwise = TRUE,
+                                               method = "DTW2",
+                                               window.type = "slantedband",
+                                               window.size = window.size))
+
+               indD <- indNN + singleIndexing
+               d[indD[indNew]] <- dSub
+          }
 
           new.indNN <- apply(d, 2, which.min)
      }